fix(android): fix onClickWidget callback handler
- rename channel
- refactor methodhandler

resolves #10
